refactor: 패키지 구조 개선 & 예약 조회 응답 필드 추가
