+++ src/main/java/roomescape/dto/reservation/ReservationResponseDto.java
@@ -3,19 +3,20 @@ package roomescape.dto.reservation;
 import java.time.format.DateTimeFormatter;
 import roomescape.domain.Reservation;
 import roomescape.domain.Theme;
+import roomescape.dto.reservationTime.ReservationTimeResponseDto;
 
 public record ReservationResponseDto(
         Long id,
         String name,
         String date,
-        String time,
+        ReservationTimeResponseDto time,
         Theme theme
 ) {
 
     public static ReservationResponseDto from(Reservation reservation) {
         String name = reservation.getName().value();
         String date = reservation.getDate().format(DateTimeFormatter.ISO_LOCAL_DATE);
-        String time = reservation.getTime().getStartAt().format(DateTimeFormatter.ofPattern("HH:mm"));
+        ReservationTimeResponseDto time = ReservationTimeResponseDto.from(reservation.getTime());
 
         return new ReservationResponseDto(
             reservation.getId(),
